Update README.md
Project Title: *Secure Remote Access Setup and Configuration for Virtual Machines

Duration: 3 weeks

*Objective:*
The objective of this project is to enable you to understand, configure, and secure remote access to virtual machines (VMs) running Windows operating systems. 
This group will be responsible for setting up a VM, ensuring it is securely accessible from a physical computer, and documenting the process and security measures.